rag/vector_store.py
import faiss
import numpy as np

class VectorStore:

    def __init__(self, dim):

        self.index = faiss.IndexFlatL2(dim)

        self.texts = []

    def add(self, embeddings, texts):

        self.index.add(np.array(embeddings))

        self.texts.extend(texts)

    def search(self, embedding, k=5):

        D,I = self.index.search(np.array([embedding]),k)

        results = []

        for idx in I[0]:

            results.append(self.texts[idx])

        return results