requirements.txt
flask
requests
feedparser
sentence-transformers
faiss-cpu
transformers
torch
numpy
ipykernel
